[dashboard] Merge column and column heading components mostly into one file
