--- app/gui/src/dashboard/pages/dashboard/components/column/PlaceholderColumn.tsx
@@ -1,6 +1,0 @@
-/** @file A placeholder component for columns which do not yet have corresponding data to display. */
-
-/** A placeholder component for columns which do not yet have corresponding data to display. */
-export default function PlaceholderColumn() {
-  return <></>
-}
