--- app/gui/src/dashboard/pages/dashboard/components/columnHeading/PathColumnHeading.tsx
@@ -1,36 +1,0 @@
-/** @file A heading for the "Path" column. */
-import { Button } from '#/components/Button'
-import { Text } from '#/components/Text'
-import { useEventCallback } from '#/hooks/eventCallbackHooks'
-import { useText } from '$/providers/react'
-import type { AssetColumnHeadingProps } from '../column'
-import { Column } from '../column/columnUtils'
-
-/** A heading for the "Path" column. */
-export default function PathColumnHeading(props: AssetColumnHeadingProps) {
-  const { hideColumn } = props
-
-  const { getText } = useText()
-
-  const hideThisColumn = useEventCallback(() => {
-    hideColumn(Column.path)
-  })
-
-  return (
-    <div
-      className="isolate flex h-table-row w-full items-center gap-2"
-      data-testid="path-column-heading"
-    >
-      <Button
-        variant="icon"
-        icon="folder"
-        aria-label={getText('pathColumnHide')}
-        tooltip={false}
-        onPress={hideThisColumn}
-      />
-      <Text weight="bold" truncate="1" color="custom">
-        {getText('pathColumnName')}
-      </Text>
-    </div>
-  )
-}
